+++ src/game/time.cpp
@@ -1,9 +1,9 @@
 #include "time.h"
 
-void time::update_time(float current_frame_time)
+void time::update_time(double current_frame_time)
 {
-    delta_time = current_frame_time - last_frame_time;
-    last_frame_time = current_frame_time;
+    delta_time = static_cast<float>(current_frame_time) - last_frame_time;
+    last_frame_time = static_cast<float>(current_frame_time);
 }
 
 float time::get_delta_time()
